requirements.txt
argparse
torchvision==0.14.1
matplotlib
pillow
librosa